// src/app/modules/result/models/term-list-entry.ts
export class TermListEntry {


    CAMPUS_CODE: string;
    USER_CODE : string;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SUBJECT_CODE: string;
    // SECTION_CODE: string;
    TERM_ID:number;
    TERM_DESCRIPTION: any[];
    TERM_DESCRIPTION_BANGLA: any[];
    START_DATE: any[];
    END_DATE: any[];
    SERIAL: any[];


}




export class IcardstudentinfoEntry {


    CAMPUS_CODE: string;
    USER_CODE : string;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SHIFT_CODE: string;
    SECTION_CODE: string;
    TERM_ID:number;
    TERM_DESCRIPTION: any[];
    START_DATE: any[];
    END_DATE: any[];
    SERIAL: any[];


}

















export class YearlyTermEntryVM{

    ORG_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    VERSION_CODE: string;
    SESSION_CODE: string;
    TERM_DESCRIPTION: any[];
    START_DATE: any[];
    END_DATE: any[];
    SERIAL: any[];
}