// src/app/modules/result/models/term-list.ts
export class TermList{
    ORG_CODE: string;
    TERM_ID: number;
    TERM_DESCRIPTION: any[];

    //CAMPUS_CODE: string;
    USER_CODE : string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    //SECTION_CODE: string; 
    START_DATE: any[];
    END_DATE: any[];
    SERIAL: any[];
}