// src/app/modules/result/models/result-grade-entry.ts
export class ResultGradeEntry{

    // Row_Status: any[];
    // GRADE_ID: number;
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SECTION_CODE: string;
    SHIFT_CODE: string;
    // SRL_NO: any[];
    FROM_VALUE: any[];
    UPTO_VALUE: any[];
    LETTER_GRADE: any[];
    GRADE_POINT: any[];
    REMARKS: any[];
    User_Name: string;
    TERM_ID: string;

}




export class GradeGridView{

    // Row_Status: any[];
    // GRADE_ID: number;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    // SRL_NO: any[];
    FROM_VALUE: any[];
    UPTO_VALUE: any[];
    LETTER_GRADE: any[];
    GRADE_POINT: any[];
    REMARKS: any[];
    User_Name: string;

}

export class ResultSummaryGrid{

    // Row_Status: any[];
    // GRADE_ID: number;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    // SRL_NO: any[];
   

}




export class AcademicCalendarVM{
    TITLE:any[];
    REASON:any[];
    START_DATE:any[];
    END_DATE:any[];



}
export class AcademicCalendar{



    ORG_CODE:string;
    TITLE:any;
    REASON:any
    START_DATE:any;
    END_DATE:any
    START_TIME:any;
    END_TIME:any
    USER_CODE:string;


}
export class ResultGradeEntryVM{

    // Row_Status: any[];
    // GRADE_ID: number;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    // SRL_NO: any[];
    FROM_VALUE: any[];
    UPTO_VALUE: any[];
    LETTER_GRADE: any[];
    GRADE_POINT: any[];
    REMARKS: any[];
    User_Name: string;

}